TechnicalResearch/github/aimarket-github/llm-router/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.57.4
python-dotenv==1.0.1
pydantic==2.10.3
requests==2.32.3
